Convert ram to be byte adressable
Memory is 16-bit addressable but should be byte addressable according to
the spec, http://devernay.free.fr/hacks/chip8/C8TECH10.HTM#00E0

+++ attar/machinedef.nim
@@ -1,5 +1,5 @@
 type chip8* = object
-  ram*: array[4096, uint16]
+  ram*: array[4096, uint8]
   pc*: uint16
   variables*: array[16, uint8]
   stack*: seq[uint16]
